+++ src/stack.c
@@ -118,7 +118,7 @@ int stack_test(void)
 
 	void *data = NULL;
 	stack_pop(s, &data);
-	printf("%d\n", data);
+	printf("%ld\n", (unsigned long)data);
 	if (s == NULL) {
 		printf("unbelievable!\n");
 		return 0;
@@ -126,9 +126,9 @@ int stack_test(void)
 	stack_show(s, show_int);
 
 	stack_pop(s, &data);
-	printf("%d\n", data);
+	printf("%ld\n", (unsigned long)data);
 	stack_pop(s, &data);
-	printf("%d\n", data);
+	printf("%ld\n", (unsigned long)data);
 
 	return 0;
 }
